+++ chandra_bot/__init__.py
@@ -1,0 +1,7 @@
+version = "0.0.1"
+
+from .chandra_bot import ChandraBot
+
+__all__ = [
+    "ChandraBot",
+]
