--- src/__init__.py
@@ -1,1 +1,0 @@
-version = "0.0.1"
